Use both distance calculations in get_stagger_adjust to determine if the source and next_step locations are equally distant from the destination.
When they are equally distant, the path planner will select the other
one as the next planned step, while only charging 1 move for the
near-zero calculation.

In some cases, a deer could spend 300 cycles of planning
and movement to flee during one turn. Zombies that are waiting on an
outer fringe of a group that are fighting can cause lag while they
attempt to get closer, because they only move back and forth over two
adjacent spaces to burn their movement points.